+++ CHANGELOG.md
@@ -14,6 +14,8 @@ See [`UPGRADE.md`](./UPGRADE.md) for additional help when upgrading to newer ver
 ### Changed
 
 * Dropped support for `RUSTLER_NIF_VERSION`
+* Deprecated `:rustler_crates` project configuration
+* Mark `use Rustler` module configuration as compile-time
 
 ## [0.29.0] - 2023-06-22